Refactor : PurchasedStock
